+++ src/Network/Network/NetworkWatcher/FlowLog/NewAzNetworkWatcherFlowLogCommand.cs
@@ -106,6 +106,12 @@ namespace Microsoft.Azure.Commands.Network
         public string StorageId { get; set; }
 
         [Parameter(
+            Mandatory = false,
+            HelpMessage = "Optional field to filter network traffic logs.")]
+        [ValidateNotNullOrEmpty]
+        public string EnabledFilteringCriteria { get; set; }
+
+        [Parameter(
             Mandatory = true,
             HelpMessage = "Flag to enable/disable flow logging.")]
         [ValidateNotNullOrEmpty]
@@ -244,7 +250,7 @@ namespace Microsoft.Azure.Commands.Network
 
         private PSFlowLogResource CreateFlowLog()
         {
-            this.ValidateFlowLogParameters(this.TargetResourceId, this.StorageId, this.FormatVersion, this.FormatType, this.EnableTrafficAnalytics == true,
+            this.ValidateFlowLogParameters(this.TargetResourceId, this.StorageId, this.EnabledFilteringCriteria, this.FormatVersion, this.FormatType, this.EnableTrafficAnalytics == true,
                 this.TrafficAnalyticsWorkspaceId, this.TrafficAnalyticsInterval, this.RetentionPolicyDays, this.UserAssignedIdentityId);
 
             MNM.FlowLog flowLogParameters = GetFlowLogParametersFromRequest();
@@ -263,6 +269,7 @@ namespace Microsoft.Azure.Commands.Network
                 TargetResourceId = this.TargetResourceId,
                 StorageId = this.StorageId,
                 Enabled = this.Enabled,
+                EnabledFilteringCriteria = this.EnabledFilteringCriteria ?? "",
                 Tags = TagsConversionHelper.CreateTagDictionary(this.Tag, validate: true)
             };