+++ src/Network/Network/NetworkWatcher/FlowLog/SetAzNetworkWatcherFlowLogCommand.cs
@@ -209,6 +209,41 @@ namespace Microsoft.Azure.Commands.Network
         public string StorageId { get; set; }
 
         [Parameter(
+            Mandatory = false,
+            HelpMessage = "Optional field to filter network traffic logs.",
+            ParameterSetName = SetByResource)]
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "Optional field to filter network traffic logs.",
+            ParameterSetName = SetByResourceWithTA)]
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "Optional field to filter network traffic logs.",
+            ParameterSetName = SetByName)]
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "Optional field to filter network traffic logs.",
+            ParameterSetName = SetByNameWithTA)]
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "Optional field to filter network traffic logs.",
+            ParameterSetName = SetByLocation)]
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "Optional field to filter network traffic logs.",
+            ParameterSetName = SetByLocationWithTA)]
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "Optional field to filter network traffic logs.",
+            ParameterSetName = SetByResourceId)]
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "Optional field to filter network traffic logs.",
+            ParameterSetName = SetByResourceIdWithTA)]
+        [ValidateNotNullOrEmpty]
+        public string EnabledFilteringCriteria { get; set; }
+
+        [Parameter(
             Mandatory = true,
             HelpMessage = "Flag to enable/disable flow logging.",
             ParameterSetName = SetByResource)]
@@ -595,6 +630,7 @@ namespace Microsoft.Azure.Commands.Network
             this.TargetResourceId = this.InputObject.TargetResourceId;
             this.StorageId = this.InputObject.StorageId;
             this.Enabled = this.InputObject.Enabled ?? true;
+            this.EnabledFilteringCriteria = this.InputObject.EnabledFilteringCriteria ?? "";
             if (this.InputObject.FlowAnalyticsConfiguration?.NetworkWatcherFlowAnalyticsConfiguration != null)
             {
                 this.EnableTrafficAnalytics = this.InputObject.FlowAnalyticsConfiguration.NetworkWatcherFlowAnalyticsConfiguration.Enabled == true;
@@ -617,7 +653,7 @@ namespace Microsoft.Azure.Commands.Network
 
         private PSFlowLogResource CreateFlowLog()
         {
-            this.ValidateFlowLogParameters(this.TargetResourceId, this.StorageId, this.FormatVersion, this.FormatType, this.EnableTrafficAnalytics == true,
+            this.ValidateFlowLogParameters(this.TargetResourceId, this.StorageId, this.EnabledFilteringCriteria, this.FormatVersion, this.FormatType, this.EnableTrafficAnalytics == true,
                 this.TrafficAnalyticsWorkspaceId, this.TrafficAnalyticsInterval, this.RetentionPolicyDays, this.UserAssignedIdentityId);
 
             MNM.FlowLog flowLogParameters = GetFlowLogParametersFromRequest();
@@ -636,6 +672,7 @@ namespace Microsoft.Azure.Commands.Network
                 TargetResourceId = this.TargetResourceId,
                 StorageId = this.StorageId,
                 Enabled = this.Enabled,
+                EnabledFilteringCriteria = this.EnabledFilteringCriteria ?? "",
                 Tags = TagsConversionHelper.CreateTagDictionary(this.Tag, validate: true)
             };